@@ -1,31 +1,131 @@
 //
 //  ChatResult.swift
-//  
+//
 //
 //  Created by Sergii Kryvoblotskyi on 02/04/2023.
 //
 
 import Foundation
 
-public struct ChatResult: Codable, Equatable {
+public struct ChatResult: Codable, Equatable, Sendable {
 
-    public struct Choice: Codable, Equatable {
-        public typealias ChatCompletionMessage = ChatQuery.ChatCompletionMessageParam
-
+    public struct Choice: Codable, Equatable, Sendable {
         /// The index of the choice in the list of choices.
         public let index: Int
         /// Log probability information for the choice.
         public let logprobs: Self.ChoiceLogprobs?
         /// A chat completion message generated by the model.
-        public let message: Self.ChatCompletionMessage
+        public let message: Self.Message
         /// The reason the model stopped generating tokens. This will be stop if the model hit a natural stop point or a provided stop sequence, length if the maximum number of tokens specified in the request was reached, content_filter if content was omitted due to a flag from our content filters, tool_calls if the model called a tool, or function_call (deprecated) if the model called a function.
-        public let finishReason: String?
+        public let finishReason: String
+        
+        public struct Message: Codable, Equatable, Sendable {
+            /// The contents of the message.
+            public let content: String?
+            /// The refusal message generated by the model.
+            public let refusal: String?
+            /// The role of the author of this message.
+            public let role: String
+            
+            /// Annotations for the message, when applicable, as when using the web search tool.
+            /// Web search tool: https://platform.openai.com/docs/guides/tools-web-search?api-mode=chat
+            ///
+            /// This field is declared non-optional in OpenAI API reference, but to not make breaking changes for our users that use this SDK for other providers like DeepSeek, made it optional for now
+            /// See https://github.com/MacPaw/OpenAI/issues/293
+            public let annotations: [Annotation]?
+            /// If the audio output modality is requested, this object contains data about the audio response from the model.
+            /// Learn more: https://platform.openai.com/docs/guides/audio
+            public let audio: Audio?
+            
+            /// The tool calls generated by the model, such as function calls.
+            ///
+            /// This field is declared not-optional in OpenAI API reference. But it seems that under some conditions it may not be present in a response.
+            /// See https://github.com/MacPaw/OpenAI/issues/293
+            public let toolCalls: [ChatQuery.ChatCompletionMessageParam.AssistantMessageParam.ToolCallParam]?
+            
+            /// Following are fields that are not part of OpenAI but are present in responses from other providers
+            
+            /// Value for `reasoning` field in response.
+            ///
+            /// Provided by:
+            /// - Gemini (in OpenAI compatibility mode)
+            ///   https://github.com/MacPaw/OpenAI/issues/283#issuecomment-2711396735
+            /// - OpenRouter
+            internal let _reasoning: String?
 
-        public struct ChoiceLogprobs: Codable, Equatable {
+            /// Value for `reasoning_content` field.
+            ///
+            /// Provided by:
+            /// - Deepseek
+            ///   https://api-docs.deepseek.com/api/create-chat-completion#responses
+            internal let _reasoningContent: String?
+
+            /// Reasoning content.
+            ///
+            /// Supported response fields:
+            /// - `reasoning` (Gemini, OpenRouter)
+            /// - `reasoning_content` (Deepseek)
+            public var reasoning: String? {
+                _reasoning ?? _reasoningContent
+            }
+
+            public enum CodingKeys: String, CodingKey {
+                case content, refusal, role, annotations, audio, toolCalls = "tool_calls"
+                case _reasoning = "reasoning"
+                case _reasoningContent = "reasoning_content"
+            }
+            
+            public struct Annotation: Codable, Equatable, Sendable {
+                /// The type of the URL citation. Always `url_citation`.
+                let type: String
+                /// A URL citation when using web search.
+                let urlCitation: URLCitation
+                
+                public enum CodingKeys: String, CodingKey {
+                    case type, urlCitation = "url_citation"
+                }
+                
+                public struct URLCitation: Codable, Equatable, Sendable {
+                    /// The index of the last character of the URL citation in the message.
+                    let endIndex: Int
+                    /// The index of the first character of the URL citation in the message.
+                    let startIndex: Int
+                    /// The title of the web resource.
+                    let title: String
+                    /// The URL of the web resource.
+                    let url: String
+                    
+                    public enum CodingKeys: String, CodingKey {
+                        case endIndex = "end_index"
+                        case startIndex = "start_index"
+                        case title, url
+                    }
+                }
+            }
+            
+            public struct Audio: Codable, Equatable, Sendable {
+                /// Base64 encoded audio bytes generated by the model, in the format specified in the request.
+                public let data: String
+                /// The Unix timestamp (in seconds) for when this audio response will no longer be accessible on the server for use in multi-turn conversations.
+                public let expiresAt: Int
+                /// Unique identifier for this audio response.
+                public let id: String
+                /// Transcript of the audio generated by the model.
+                public let transcript: String
+                
+                public enum CodingKeys: String, CodingKey {
+                    case data
+                    case expiresAt = "expires_at"
+                    case id, transcript
+                }
+            }
+        }
+
+        public struct ChoiceLogprobs: Codable, Equatable, Sendable {
 
             public let content: [Self.ChatCompletionTokenLogprob]?
 
-            public struct ChatCompletionTokenLogprob: Codable, Equatable {
+            public struct ChatCompletionTokenLogprob: Codable, Equatable, Sendable {
 
                 /// The token.
                 public let token: String
@@ -40,7 +140,7 @@ public struct ChatResult: Codable, Equatable {
                 /// In rare cases, there may be fewer than the number of requested `top_logprobs` returned.
                 public let topLogprobs: [TopLogprob]
 
-                public struct TopLogprob: Codable, Equatable {
+                public struct TopLogprob: Codable, Equatable, Sendable {
 
                     /// The token.
                     public let token: String
@@ -67,16 +167,17 @@ public struct ChatResult: Codable, Equatable {
             case finishReason = "finish_reason"
         }
 
-        public enum FinishReason: String, Codable, Equatable {
+        public enum FinishReason: String, Codable, Equatable, Sendable {
             case stop
             case length
             case toolCalls = "tool_calls"
             case contentFilter = "content_filter"
             case functionCall = "function_call"
+            case error
         }
     }
 
-    public struct CompletionUsage: Codable, Equatable {
+    public struct CompletionUsage: Codable, Equatable, Sendable {
 
         /// Number of tokens in the generated completion.
         public let completionTokens: Int
@@ -94,19 +195,28 @@ public struct ChatResult: Codable, Equatable {
 
     /// A unique identifier for the chat completion.
     public let id: String
-    /// The object type, which is always chat.completion.
-    public let object: String
     /// The Unix timestamp (in seconds) of when the chat completion was created.
-    public let created: TimeInterval
+    public let created: Int
     /// The model used for the chat completion.
     public let model: String
+    /// The object type, which is always `chat.completion`.
+    public let object: String
+    /// The service tier used for processing the request.
+    public let serviceTier: String?
+    /// This fingerprint represents the backend configuration that the model runs with.
+    /// Can be used in conjunction with the seed request parameter to understand when backend changes have been made that might impact determinism.
+    public let systemFingerprint: String
     /// A list of chat completion choices. Can be more than one if n is greater than 1.
     public let choices: [Choice]
     /// Usage statistics for the completion request.
     public let usage: Self.CompletionUsage?
-    /// This fingerprint represents the backend configuration that the model runs with.
-    /// Can be used in conjunction with the seed request parameter to understand when backend changes have been made that might impact determinism.
-    public let systemFingerprint: String?
+    
+    /// Following are fields that are not part of OpenAI API Reference, but are present in responses from other providers
+    ///
+    /// Perplexity
+    /// Citations for the generated answer.
+    /// https://docs.perplexity.ai/api-reference/chat-completions#response-citations
+    public let citations: [String]?
 
     public enum CodingKeys: String, CodingKey {
         case id
@@ -114,8 +224,36 @@ public struct ChatResult: Codable, Equatable {
         case created
         case model
         case choices
-        case usage
+        case serviceTier = "service_tier"
         case systemFingerprint = "system_fingerprint"
+        case usage
+        case citations
+    }
+    
+    init(id: String, created: Int, model: String, object: String, serviceTier: String? = nil, systemFingerprint: String, choices: [Choice], usage: Self.CompletionUsage? = nil, citations: [String]? = nil) {
+        self.id = id
+        self.created = created
+        self.model = model
+        self.object = object
+        self.serviceTier = serviceTier
+        self.systemFingerprint = systemFingerprint
+        self.choices = choices
+        self.usage = usage
+        self.citations = citations
+    }
+    
+    public init(from decoder: any Decoder) throws {
+        let container = try decoder.container(keyedBy: CodingKeys.self)
+        let parsingOptions = decoder.userInfo[.parsingOptions] as? ParsingOptions ?? []
+        self.id = try container.decodeString(forKey: .id, parsingOptions: parsingOptions)
+        self.object = try container.decodeString(forKey: .object, parsingOptions: parsingOptions)
+        self.created = try container.decode(Int.self, forKey: .created)
+        self.model = try container.decodeString(forKey: .model, parsingOptions: parsingOptions)
+        self.choices = try container.decode([ChatResult.Choice].self, forKey: .choices)
+        self.serviceTier = try container.decodeIfPresent(String.self, forKey: .serviceTier)
+        self.systemFingerprint = try container.decodeString(forKey: .systemFingerprint, parsingOptions: parsingOptions)
+        self.usage = try container.decodeIfPresent(ChatResult.CompletionUsage.self, forKey: .usage)
+        self.citations = try container.decodeIfPresent([String].self, forKey: .citations)
     }
 }
 
@@ -126,6 +264,8 @@ extension ChatQuery.ChatCompletionMessageParam {
         switch try messageContainer.decode(Role.self, forKey: .role) {
         case .system:
             self = try .system(.init(from: decoder))
+        case .developer:
+            self = try .developer(.init(from: decoder))
         case .user:
             self = try .user(.init(from: decoder))
         case .assistant:
@@ -136,7 +276,7 @@ extension ChatQuery.ChatCompletionMessageParam {
     }
 }
 
-extension ChatQuery.ChatCompletionMessageParam.ChatCompletionUserMessageParam.Content {
+extension ChatQuery.ChatCompletionMessageParam.UserMessageParam.Content {
     public init(from decoder: Decoder) throws {
         let container = try decoder.singleValueContainer()
         do {
