@@ -8,13 +8,13 @@
 import Foundation
 
 /// Returns a list of image objects.
-public struct ImagesResult: Codable, Equatable {
+public struct ImagesResult: Codable, Equatable, Sendable {
 
     public let created: TimeInterval
     public let data: [Self.Image]
 
     /// Represents the url or the content of an image generated by the OpenAI API.
-    public struct Image: Codable, Equatable {
+    public struct Image: Codable, Equatable, Sendable {
 
         /// The base64-encoded JSON of the generated image, if response_format is b64_json
         public let b64Json: String?
