@@ -7,7 +7,7 @@
 
 import Foundation
 
-public struct ModerationsQuery: Codable {
+public struct ModerationsQuery: Codable, Sendable {
     /// The input text to classify.
     public let input: String
     /// ID of the model to use.
