@@ -7,12 +7,11 @@
 
 import Foundation
 
-@available(iOS 13.0, *)
-public struct ModerationsResult: Codable, Equatable {
+public struct ModerationsResult: Codable, Equatable, Sendable {
     
-    public struct Moderation: Codable, Equatable {
+    public struct Moderation: Codable, Equatable, Sendable {
         
-        public struct Categories: Codable, Equatable, Sequence {
+        public struct Categories: Codable, Equatable, Sequence, Sendable {
 
             /// Content that expresses, incites, or promotes harassing language towards any target.
             public let harassment: Bool
@@ -58,7 +57,7 @@ public struct ModerationsResult: Codable, Equatable {
             }
         }
 
-        public struct CategoryScores: Codable, Equatable, Sequence {
+        public struct CategoryScores: Codable, Equatable, Sequence, Sendable {
 
             /// Content that expresses, incites, or promotes harassing language towards any target.
             public let harassment: Double
@@ -123,5 +122,4 @@ public struct ModerationsResult: Codable, Equatable {
     public let results: [Self.Moderation]
 }
 
-@available(iOS 13.0, *)
 extension ModerationsResult: Identifiable {}
