@@ -11,16 +11,23 @@ import FoundationNetworking
 #endif
 @testable import OpenAI
 
-class DataTaskMock: URLSessionDataTaskProtocol {
+class DataTaskMock: URLSessionDataTaskProtocol, @unchecked Sendable {
     
     var data: Data?
     var response: URLResponse?
     var error: Error?
-    
+    var urlError: URLError? // Needed for mocking combine dataTaskPublisher
+    var originalRequest: URLRequest?
+
     var completion: ((Data?, URLResponse?, Error?) -> Void)?
     
     func resume() {
         completion?(data, response, error)
+    }
+    
+    var cancelCallCount = 0
+    func cancel() {
+        cancelCallCount += 1
     }
 }
 
