@@ -13,7 +13,9 @@ struct APIProvidedView: View {
     @Binding var apiKey: String
     @StateObject var chatStore: ChatStore
     @StateObject var imageStore: ImageStore
+    @StateObject var assistantStore: AssistantStore
     @StateObject var miscStore: MiscStore
+
     @State var isShowingAPIConfigModal: Bool = true
 
     @Environment(\.idProviderValue) var idProvider
@@ -24,20 +26,28 @@ struct APIProvidedView: View {
         idProvider: @escaping () -> String
     ) {
         self._apiKey = apiKey
+        
+        let client = APIProvidedView.makeClient(apiKey: apiKey.wrappedValue)
         self._chatStore = StateObject(
             wrappedValue: ChatStore(
-                openAIClient: OpenAI(apiToken: apiKey.wrappedValue),
+                openAIClient: client,
                 idProvider: idProvider
             )
         )
         self._imageStore = StateObject(
             wrappedValue: ImageStore(
-                openAIClient: OpenAI(apiToken: apiKey.wrappedValue)
+                openAIClient: client
+            )
+        )
+        self._assistantStore = StateObject(
+            wrappedValue: AssistantStore(
+                openAIClient: client,
+                idProvider: idProvider
             )
         )
         self._miscStore = StateObject(
             wrappedValue: MiscStore(
-                openAIClient: OpenAI(apiToken: apiKey.wrappedValue)
+                openAIClient: client
             )
         )
     }
@@ -46,13 +56,19 @@ struct APIProvidedView: View {
         ContentView(
             chatStore: chatStore,
             imageStore: imageStore,
+            assistantStore: assistantStore,
             miscStore: miscStore
         )
         .onChange(of: apiKey) { newApiKey in
-            let client = OpenAI(apiToken: newApiKey)
+            let client = APIProvidedView.makeClient(apiKey: newApiKey)
             chatStore.openAIClient = client
             imageStore.openAIClient = client
+            assistantStore.openAIClient = client
             miscStore.openAIClient = client
         }
+    }
+    
+    private static func makeClient(apiKey: String) -> OpenAIProtocol {
+        OpenAI(apiToken: apiKey)
     }
 }
