@@ -19,9 +19,9 @@ public struct ModerationChatView: View {
     
     public var body: some View {
         DetailView(
-            conversation: store.moderationConversation,
+            availableAssistants: [], conversation: store.moderationConversation, 
             error: store.moderationConversationError,
-            sendMessage: { message, _ in
+            sendMessage: { message, _, _ in
                 Task {
                     await store.sendModerationMessage(
                         Message(
@@ -32,7 +32,7 @@ public struct ModerationChatView: View {
                         )
                     )
                 }
-            }
+            }, isSendingMessage: Binding.constant(false)
         )
     }
 }
