@@ -11,8 +11,23 @@ import FoundationNetworking
 #endif
 
 protocol URLRequestBuildable {
-    
-    associatedtype ResultType
-    
-    func build(token: String, organizationIdentifier: String?, timeoutInterval: TimeInterval) throws -> URLRequest
+    func build(
+        token: String?,
+        organizationIdentifier: String?,
+        timeoutInterval: TimeInterval,
+        customHeaders: [String: String]
+    ) throws -> URLRequest
+}
+
+extension URLRequestBuildable {
+    func build(
+        configuration: OpenAI.Configuration
+    ) throws -> URLRequest {
+        try build(
+            token: configuration.token,
+            organizationIdentifier: configuration.organizationIdentifier,
+            timeoutInterval: configuration.timeoutInterval,
+            customHeaders: configuration.customHeaders
+        )
+    }
 }
