@@ -10,13 +10,44 @@ public typealias Model = String
 
 public extension Model {
     // Chat Completion
+    
+    // o3 series
+    /// `o3-mini`: small reasoning model providing high intelligence at the same cost and latency targets of o1-mini.
+    static let o3_mini = "o3-mini"
+    
+    // o1 series
+    // The o1 series of models are trained with reinforcement learning to perform complex reasoning. o1 models think before they answer, producing a long internal chain of thought before responding to the user.
+    
+    /// `o1`: reasoning model designed to solve hard problems across domains
+    static let o1 = "o1"
+    
+    /// `o1-mini`: fast and affordable reasoning model for specialized tasks
+    static let o1_mini = "o1-mini"
+
+    // GPT-4.1
+
+    /// `gpt-4.1` Smartest model for complex tasks
+    static let gpt4_1 = "gpt-4.1"
+
+    /// `gpt-4.1-mini` Affordable model balancing speed and intelligence
+    static let gpt4_1_mini = "gpt-4.1-mini"
+
+    /// `gpt-4.1-nano` Fastest, most cost-effective model for low-latency tasks
+    static let gpt4_1_nano = "gpt-4.1-nano"
+
     // GPT-4
 
     /// `gpt-4o`, currently the most advanced, multimodal flagship model that's cheaper and faster than GPT-4 Turbo.
     static let gpt4_o = "gpt-4o"
-
+    
     /// `gpt-4o-mini`, currently the most affordable and intelligent model for fast and lightweight requests.
     static let gpt4_o_mini = "gpt-4o-mini"
+
+    /// `gpt-4o-audio-preview`, this is a preview release of the GPT-4o Audio models. These models accept audio inputs and outputs, and can be used in the Chat Completions REST API.
+    static let gpt_4o_audio_preview = "gpt-4o-audio-preview"
+
+    /// `gpt-4o-mini-audio-preview`, this is a preview release of the smaller GPT-4o Audio mini model. It's designed to input audio or create audio outputs via the REST API.
+    static let gpt_4o_mini_audio_preview = "gpt-4o-mini-audio-preview"
 
     /// `gpt-4-turbo`, The latest GPT-4 Turbo model with vision capabilities. Vision requests can now use JSON mode and function calling and more. Context window: 128,000 tokens
     static let gpt4_turbo = "gpt-4-turbo"
@@ -54,6 +85,9 @@ public extension Model {
     /// Snapshot of `gpt-4-32k` from March 14th 2023. Unlike `gpt-4-32k`, this model will not receive updates, and will only be supported for a three month period ending on June 14th 2023.
     @available(*, deprecated, message: "Please upgrade to the newer model")
     static let gpt4_32k_0314 = "gpt-4-32k-0314"
+    
+    /// Research preview of GPT-4.5, largest and most capable GPT model yet. As for Feb 28 2025 points to gpt-4.5-preview-2025-02-27
+    static let gpt4_5_preview = "gpt-4.5-preview"
 
     // GPT-3.5
     
@@ -80,35 +114,23 @@ public extension Model {
     
     /// Snapshot of `gpt-3.5-turbo-16k` from June 13th 2023. Unlike `gpt-3.5-turbo-16k`, this model will not receive updates, and will be deprecated 3 months after a new version is released.
     static let gpt3_5Turbo_16k_0613 = "gpt-3.5-turbo-16k-0613"
-
-    // Completions
-    
-    /// Can do any language task with better quality, longer output, and consistent instruction-following than the curie, babbage, or ada models. Also supports inserting completions within text.
-    static let textDavinci_003 = "text-davinci-003"
-    /// Similar capabilities to text-davinci-003 but trained with supervised fine-tuning instead of reinforcement learning.
-    static let textDavinci_002 = "text-davinci-002"
-    /// Very capable, faster and lower cost than Davinci.
-    static let textCurie = "text-curie-001"
-    /// Capable of straightforward tasks, very fast, and lower cost.
-    static let textBabbage = "text-babbage-001"
-    /// Capable of very simple tasks, usually the fastest model in the GPT-3 series, and lowest cost.
-    static let textAda = "text-ada-001"
-    
-    // Edits
-    
-    static let textDavinci_001 = "text-davinci-001"
-    static let codeDavinciEdit_001 = "code-davinci-edit-001"
     
     // Speech
-    
-    /// The latest text to speech model, optimized for speed.
+    /// A text-to-speech model built on GPT-4o mini, a fast and powerful language model.
+    static let gpt_4o_mini_tts = "gpt-4o-mini-tts"
+    /// tts-1 model is optimized for realtime text-to-speech use cases
     static let tts_1 = "tts-1"
-    /// The latest text to speech model, optimized for quality.
+    /// tts-1-hd model is optimized for high quality text-to-speech use cases.
     static let tts_1_hd = "tts-1-hd"
     
     // Transcriptions / Translations
-    
+
+    /// Whisper is a general-purpose speech recognition model, trained on a large dataset of diverse audio. You can also use it as a multitask model to perform multilingual speech recognition as well as speech translation and language identification.
     static let whisper_1 = "whisper-1"
+    /// GPT-4o Transcribe is a speech-to-text model that uses GPT-4o to transcribe audio. It offers improvements to word error rate and better language recognition and accuracy compared to original Whisper models. Use it for more accurate transcripts.
+    static let gpt_4o_transcribe = "gpt-4o-transcribe"
+    /// GPT-4o mini Transcribe is a speech-to-text model that uses GPT-4o mini to transcribe audio. It offers improvements to word error rate and better language recognition and accuracy compared to original Whisper models. Use it for more accurate transcripts.
+    static let gpt_4o_mini_transcribe = "gpt-4o-mini-transcribe"
 
     // Image Generation
     static let dall_e_2 = "dall-e-2"
